can filter data based on the active schoolyear. need adjustments on the other features

+++ database/migrations/2025_02_22_130625_add_column_to_user_schoolyear_table.php
@@ -12,8 +12,15 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('user_schoolyear', function (Blueprint $table) {
-            //fk for users and schoolyears
-            $table->foreignId('schoolyear_id')->constrained('schoolyears')->onDelete('cascade');
+            // //fk for users and schoolyears
+            // $table->foreignId('user_id')->constrained('users')->onDelete('cascade');
+            // $table->foreignId('schoolyear_id')->constrained('schoolyears')->onDelete('cascade');
+            if (!Schema::hasColumn('user_schoolyear', 'user_id')) {
+                $table->foreignId('user_id')->constrained('users')->onDelete('cascade');
+            }
+            if (!Schema::hasColumn('user_schoolyear', 'schoolyear_id')) {
+                $table->foreignId('schoolyear_id')->constrained('schoolyears')->onDelete('cascade');
+            }
         });
     }
 
@@ -25,9 +32,11 @@ return new class extends Migration
         Schema::table('user_schoolyear', function (Blueprint $table) {
             
             //Drop foreign key constraints            $table->dropForeign(['schoolyear_id']);
-
+            $table->dropForeign(['user_id']);
+            $table->dropForeign(['schoolyear_id']);
             // Drop columns
             $table->dropColumn(['schoolyear_id']);
+            $table->dropColumn(['user_id']);
         });
     }
 };
